idx type Long으로 변경 support 및 common #41

+++ src/main/java/com/tsp/new_tsp_front/api/common/domain/CommonImageDTO.java
@@ -16,10 +16,10 @@ import lombok.experimental.SuperBuilder;
 @ApiModel
 public class CommonImageDTO {
     @ApiModelProperty(value = "파일 IDX", required = true, hidden = true)
-    private Integer idx;
+    private Long idx;
 
     @ApiModelProperty(value = "분야 IDX", required = true, hidden = true)
-    private Integer typeIdx;
+    private Long typeIdx;
 
     @ApiModelProperty(value = "분야명", required = true, hidden = true)
     private String typeName;
